Added dragdrop feature, organized files in components, route to transcript
