+++ mikan-vuetify/src/components.d.ts
@@ -3,34 +3,41 @@
 // Generated by unplugin-vue-components
 // Read more: https://github.com/vuejs/core/pull/3399
 // biome-ignore lint: disable
-export {}
+export {};
 
 /* prettier-ignore */
 declare module 'vue' {
   export interface GlobalComponents {
     ActionButtonsBar: typeof import('./components/ActionButtonsBar.vue')['default']
     AppFooter: typeof import('./components/AppFooter.vue')['default']
-    AudioConverter: typeof import('./components/AudioConverter.vue')['default']
-    DownloadButton: typeof import('./components/Buttons/DownloadButton.vue')['default']
-    EditButton: typeof import('./components/Buttons/EditButton.vue')['default']
+    ApproveButtonDialogue: typeof import('./components/ApproveButtonDialogue.vue')['default']
+    AudioConverter: typeof import('./components/AudioComponent/AudioConverter.vue')['default']
+    DeleteButtonDialogue: typeof import('./components/DeleteButtonDialogue.vue')['default']
+    DownloadButton: typeof import('./components/MinutesComponent/DownloadButton.vue')['default']
+    EditButton: typeof import('./components/MinutesComponent/EditButton.vue')['default']
+    EditTaskDialog: typeof import('./components/EditTaskDialog.vue')['default']
     HelloWorld: typeof import('./components/HelloWorld.vue')['default']
-    KanbanButton: typeof import('./components/Buttons/KanbanButton.vue')['default']
+    KanbanButton: typeof import('./components/MinutesComponent/KanbanButton.vue')['default']
     KanbanRow: typeof import('./components/KanbanRow.vue')['default']
     Lightbulb: typeof import('./components/Lightbulb.vue')['default']
-    MinuteDisplay: typeof import('./components/MinuteDisplay.vue')['default']
-    MinutesButton: typeof import('./components/Buttons/MinutesButton.vue')['default']
+    MinuteDisplay: typeof import('./components/MinutesComponent/MinuteDisplay.vue')['default']
+    MinutesButton: typeof import('./components/TranscriptComponent/MinutesButton.vue')['default']
     MyAppBar: typeof import('./components/MyAppBar.vue')['default']
-    Navbar: typeof import('./components/Navbar.vue')['default']
+    MyAppBar1: typeof import('./components/AudioComponent/MyAppBar1.vue')['default']
+    MyAppBar2: typeof import('./components/MinutesComponent/MyAppBar2.vue')['default']
+    MyAppBar3: typeof import('./components/TranscriptComponent/MyAppBar3.vue')['default']
+    Navbar: typeof import('./components/AudioComponent/Navbar.vue')['default']
     NavbarManagerReview: typeof import('./components/NavbarManagerReview.vue')['default']
     PopUpConfirmDelete: typeof import('./components/PopUpConfirmDelete.vue')['default']
     ProjectBoards: typeof import('./components/ProjectBoards.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
-    ShareButton: typeof import('./components/Buttons/ShareButton.vue')['default']
-    Sidebar: typeof import('./components/Sidebar.vue')['default']
+    ShareButton: typeof import('./components/MinutesComponent/ShareButton.vue')['default']
+    Sidebar: typeof import('./components/AudioComponent/Sidebar.vue')['default']
     TaskDraft: typeof import('./components/TaskDraft.vue')['default']
-    TasklistButton: typeof import('./components/Buttons/TasklistButton.vue')['default']
-    TranscriptButton: typeof import('./components/Buttons/TranscriptButton.vue')['default']
-    TranscriptDisplay: typeof import('./components/TranscriptDisplay.vue')['default']
+    TasklistButton: typeof import('./components/MinutesComponent/TasklistButton.vue')['default']
+    TranscriptButton: typeof import('./components/MinutesComponent/TranscriptButton.vue')['default']
+    TranscriptDisplay: typeof import('./components/TranscriptComponent/TranscriptDisplay.vue')['default']
+    Uploadfile: typeof import('./components/AudioComponent/Uploadfile.vue')['default']
   }
 }
