+++ mikan-vuetify/src/pages/Homepage.vue
@@ -1,16 +1,7 @@
-<template> 
-    <Sidebar />
-    <MyAppBar />  
-    <v-app>
-    <v-main>
-
-      <Lightbulb />
-    <AudioConverter />
-    </v-main></v-app>
+<template>
+  <Sidebar />
+  <MyAppBar1 />
+  <AudioConverter />
 </template>
 
-<script>
-import Sidebar from "../components/Sidebar.vue";
-import Navbar from "../components/Navbar.vue";
-import AudioConverter from "../components/AudioConverter.vue";
-</script>
+<script></script>
