--- mikan-vuetify/src/pages/Transcripts.vue
@@ -1,17 +1,0 @@
-<template>
-   
-  <Sidebar /> 
-  <MyAppBar /> 
-  <TranscriptDisplay />
-</template>
-
-<script setup>  
-</script>
-
-<style scoped>
-/* Add this style to your transcript.vue if not already global */
-.content-block {
-  background-color: #FFFFFF; /* Each section row background color */
-  box-shadow: 0 2px 4px rgba(0,0,0,0.05);
-}
-</style>
