+++ mikan-vuetify/src/pages/TranscriptsHomepage.vue
@@ -1,0 +1,15 @@
+<template>
+  <Sidebar />
+  <MyAppBar3 />
+  <TranscriptDisplay />
+</template>
+
+<script setup></script>
+
+<style scoped>
+/* Add this style to your transcript.vue if not already global */
+.content-block {
+  background-color: #ffffff; /* Each section row background color */
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
+}
+</style>
